BIG UPDATE - Design improvements, lang and others - Performance improvements - Add content, awards
I updated the lang design, when you are on a touchscreen, you can click on the language selector to appear the other language choices and click again to hide the selector and this is working too when you click outside the lang element. On common display like computers, you just need to hover the language selector like before. I renamed some classes for better reading.

I improved the performance of the page, like lazy loading for images before the final load the display of these images fastly. I added fonts fallback and the "font-display: swap" to instructs the browser to use the fallback font to display the text until the custom font has fully downloaded.

I added my letters of merit by the Cisco Networking Academy. I finished my last Netacad certification the 15/05/2020.